+++ spec/scalingo/configuration_spec.rb
@@ -3,36 +3,6 @@ require "spec_helper"
 RSpec.describe Scalingo::Configuration do
   subject { described_class.default }
 
-  describe "default_region" do
-    it "must be an existing region" do
-      expect {
-        subject.default_region = "another-region"
-      }.to raise_error(ArgumentError)
-    end
-  end
-
-  describe "regions" do
-    it "can be assigned from a hash" do
-      subject.regions = {
-        local_name: "some-url",
-      }
-
-      expect(subject.regions.local_name).to eq "some-url"
-    end
-
-    it "can be assigned from a openstruct" do
-      subject.regions = OpenStruct.new(local_name: "some-url")
-
-      expect(subject.regions.local_name).to eq "some-url"
-    end
-
-    it "raises with an argument from the wrong type" do
-      expect {
-        subject.regions = "1"
-      }.to raise_error(ArgumentError)
-    end
-  end
-
   describe "inheritance" do
     it "can inherit configuration from a parent" do
       object = described_class.new({}, subject)